assignment_3/pyproject.toml
[project]
name = "assignment-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "google-generativeai",
    "python-dotenv",
]
